python script
It is a small project about people analytics analysis. I created a script in python using pandas and numpy for merged two randons datasets about managers in general, and tableau software for create a dashboard for visualization. The main goal was answered:  What are the characteristics of a good manager?